==> 5_Method_selection_simulation.Rmd <==
---
title: "Selection of updating method for MINDACT - simulation"
author: "Mary Ann Binuya"
date: "`r Sys.Date()`"
output:
  html_document:
    toc: true
    toc_float: true
  pdf_document: default
---

# Aim
1. Sensitivity analysis of closed test procedure applied on MINDACT sub-samples

```{r setup, include=FALSE}
#Set libpath
#.libPaths(c("U:/libraries", .libPaths()))

knitr::opts_chunk$set(
  echo = TRUE,
  message = TRUE,
  warning = FALSE,
  fig.retina = 3,
  fig.path = "Output/5 Closed Test Results - simulation/")

#Load libraries
library(dplyr) #for data manipulations
library(survival) #for survival analysis
library(rms) #for survival analysis
#library(glmnet) #supports survival models
#library(caret) #doesn't support survival models
library(ggplot2) #ggplot
library(viridis) #color blind friendly palette

```

# Load data
```{r Load}
load("WS_1_Data_Prep_MINDACT.RData")

dat <- datapredimp_bc %>%
  filter(.imp!=0) %>%
  select(.imp,
         time, #Follow-up time
         eventbc, #Breast cancer death
         pirx, #Linear predictor of PREDICT BC death ("clinical risk")
         age.start, #Age at diagnosis
         size, #Tumor size
         grade, #Tumor grade
         nodes, #Number of nodes
         #er, #ER status
         pr, #Progesterone status
         her2, #HER2 status
         ki67, #Ki67 status
         generation, #Chemotherapy generation
         horm, #Hormone therapy
         traz, #Trastuzumab (anti-HER2 therapy)
         bis, #Bisphosphonates
         grisk) #Genomic risk (MammaPrint)

  rm(list=setdiff(ls(), c("dat")))
  
max(dat$.imp) #nimp
```

# Simulation
```{r Simulation}
ClosedTestCox_index <- function(data, nimp = 20, alpha_test = 0.05) {
  X <- data[, c("pirx", "age.start", "size", "grade", "nodes", "pr", "her2", "ki67", "generation", "horm", "traz", "bis")]
  X <- model.matrix( ~ ., X)[, -1]
  df <- data.frame(time = data$time, event = data$eventbc, X)
  df$wt <- 1/nimp
  
  # Method 1: Recalibration-in-the-large
  model_recal_large <- coxph(Surv(time, event) ~ offset(pirx), weights = wt, data = df)
  ll_recal_large <- as.numeric(logLik(model_recal_large))
  
  # Method 2: Recalibration
  model_recal <- coxph(Surv(time, event) ~ pirx,  weights = wt, data = df)
  ll_recal <- as.numeric(logLik(model_recal))

  # Method 3: Revision (based on pre-selection)
  model_rev <- coxph(Surv(time, event) ~ . -wt ,  weights = wt, data = df)
  ll_rev <- as.numeric(logLik(model_rev))
  df_rev <- ncol(X)

  # Calculate difference in partial log likelihoods
  dev_keep_recal_large <- 2*(ll_rev - ll_recal_large)
  dev_keep_recal <- 2*(ll_rev - ll_recal)
  
  pval1 <- (1 - pchisq(dev_keep_recal_large, df_rev))
  pval2 <- (1 - pchisq(dev_keep_recal, df_rev - 1))
  
  stats <- data.frame(round(dev_keep_recal_large,2), df_rev, round(pval1,4),
                      round(dev_keep_recal,2), df_rev - 1, round(pval2,4))
  colnames(stats) <- c("Test1_chisq", "Test1_df", "Test1_pval",
                       "Test2_chisq", "Test2_df", "Test2_pval")
  
  # See if difference in model fit is significant
  test1 <- pval1 < alpha_test
  test2 <- pval2 < alpha_test
  
  # See which model is chosen, index_test indicates the chosen model
  test_recal_large <- 1 * ((!test1))
  test_recal <- 2 * ((test1) & (!test2))
  test_rev <- 3 * ((test1) & (test2))
  index_test <- (test_recal_large + test_recal + test_rev)
  
  # Output
  model <- c("Model with updated baseline risk",
             "Recalibrated model",
             "Model revision")[index_test]
  return(model)
}

dat_with_id <- dat %>%
  filter(.imp != 0) %>%
  group_by(.imp) %>%
  mutate(id = row_number()) %>% #create id for each record
  ungroup() %>%
  select(.imp, id, time, eventbc, pirx, age.start, size, grade, nodes,
         pr, her2, ki67, generation, horm, traz, bis, grisk)

simulation <- function(patients_data, sample_sizes, num_simulations) {
  results <- data.frame(
    sample_size = integer(),
    event_size = integer(),
    closed_test_result = character())
  
  for (n in sample_sizes) {
    cat("Running simulation for sample size:", n, "\n")
    
    for (i in 1:num_simulations) {
      # Select sample IDs
      sample_ids <- sample(unique(patients_data$id), n, replace = FALSE) #without replacement
      
      sample_data <- patients_data %>%
        filter(id %in% sample_ids)
      
      event_size <- sum(sample_data$eventbc[sample_data$.imp == 1]) #same for each .imp
      
      closed_test_result <- ClosedTestCox_index(sample_data)
      
      results <- rbind(results, data.frame(
        sample_size = n,
        event_size = event_size,
        closed_test_result = closed_test_result
      ))
    }
  }
  
  return(results)
}

# Define range of sample sizes and number of simulations
sample_sizes <- seq(200, 5920, by = 100)
num_simulations <- 100 #smaller number selected for efficiency due to multiple imputation

# Run simulations
start_time <- Sys.time()
set.seed(123)  #for reproducibility
simulation_results <- simulation(dat_with_id, sample_sizes, num_simulations)
end_time <- Sys.time()
end_time - start_time

# Plot results
ggplot(simulation_results, aes(x = sample_size, y = event_size, color = closed_test_result)) +
  geom_point(alpha = 0.5) +
  scale_color_discrete(name = "Closed Test Result") +
  labs(title = "Simulation Results", x = "Sample Size", y = "Event Size") +
  theme_minimal()

ggplot(simulation_results, aes(x = sample_size, y = event_size, color = closed_test_result)) +
  geom_point(alpha = 0.5) +
  scale_color_viridis_d(name = "Closed Test Result") + #color-blind friendly
  labs(title = "Simulation Results", x = "Sample Size", y = "Event Size") +
  theme_minimal()

#save.image("WS_5_Selection_100simulationspersample.RData")
```
